backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
requests==2.32.5
python-dotenv==1.2.2
apscheduler==3.11.2
sqlalchemy==2.0.48
aiohttp==3.13.3
pydantic==2.12.5
pydantic[email]==2.12.5
shodan==1.31.0
reportlab==4.4.10
httpx==0.28.1
slowapi==0.1.9
beautifulsoup4==4.14.3
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
email-validator==2.3.0
psycopg2-binary==2.9.10
pytest==8.3.5
